requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
PyNaCl>=1.5.0
yt-dlp>=2023.11.16
aiohttp>=3.9.1
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
pillow>=10.1.0
aiosqlite>=0.19.0
colorama>=0.4.6
black>=23.12.1
isort>=5.13.2
flake8>=7.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
mypy>=1.8.0
types-requests>=2.31.0.20240125 